pyproject.toml
[tool.poetry]
name = "sisyphus-handbrake"
version = "1.0.0a1"
description = ""
authors = ["JamesTheBard <jamesthebard@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
python-box = "^7.1.1"
jsonschema = "^4.19.1"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
